+++ apps/tutorials/ospTutorialStructuredVolume.cpp
@@ -34,12 +34,48 @@ static void setIsoValue(OSPGeometry geometry, float value)
   ospRelease(isoValuesData);
 }
 
-static void setSlice(OSPGeometry geometry, float value)
+static void setSliceGeometry(OSPGeometry sliceGeometry, float sliceExtent, float slicePosition)
 {
-  vec4f plane(-1.f, 0.f, 0.f, value);
-  OSPData planesData = ospNewData(1, OSP_VEC4F, &plane);
-  ospSetData(geometry, "plane", planesData);
-  ospRelease(planesData);
+  std::vector<vec3f> vertices;
+  std::vector<int> sliceIndices;
+
+  vertices.push_back(vec3f{-slicePosition, -sliceExtent, -sliceExtent});
+  vertices.push_back(vec3f{-slicePosition, sliceExtent, -sliceExtent});
+  vertices.push_back(vec3f{-slicePosition, sliceExtent, sliceExtent});
+  vertices.push_back(vec3f{-slicePosition, -sliceExtent, sliceExtent});
+
+  int indices = vertices.size();
+  for (int i = 0; i < indices ; ++i) {
+    sliceIndices.push_back(i);
+  }
+  OSPData positionData =
+      ospNewData(vertices.size(), OSP_VEC3F, vertices.data());
+  OSPData indexData =
+      ospNewData(sliceIndices.size(), OSP_VEC4I, sliceIndices.data());
+
+  ospSetData(sliceGeometry, "vertex.position", positionData);
+  ospSetData(sliceGeometry, "index", indexData);
+
+  ospRelease(positionData);
+  ospRelease(indexData);
+}
+
+static OSPGeometricModel setSliceVolumeModel(OSPVolumetricModel volumeModel,
+  OSPGeometricModel sliceModel)
+{
+  OSPMaterial sliceMaterial = ospNewMaterial(renderer_type.c_str(), "default");
+  OSPTexture sliceTex = ospNewTexture("volume");
+  ospSetObject(sliceTex, "volume", volumeModel);
+  ospCommit(sliceTex);
+  ospSetObject(sliceMaterial, "map_Kd", sliceTex);
+  ospCommit(sliceMaterial);
+  ospSetObject(sliceModel, "material", sliceMaterial);
+  ospCommit(sliceModel);
+
+  ospRelease(sliceTex);
+  ospRelease(sliceMaterial);
+
+  return sliceModel;
 }
 
 int main(int argc, const char **argv)
@@ -71,14 +107,19 @@ int main(int argc, const char **argv)
 
   OSPTransferFunction tfn =
       ospTestingNewTransferFunction(test_data.voxelRange, "jet");
+  OSPTransferFunction stfn =
+      ospTestingNewTransferFunction(test_data.voxelRange, "grayscale");
 
   auto volumeModel = ospNewVolumetricModel(volume);
+  auto sliceVolumeModel = ospNewVolumetricModel(volume);
   ospSetObject(volumeModel, "transferFunction", tfn);
+  ospSetObject(sliceVolumeModel, "transferFunction", stfn);
   ospSetFloat(volumeModel, "samplingRate", 0.5f);
   ospCommit(volumeModel);
+  ospCommit(sliceVolumeModel);
 
   ospRelease(tfn);
-
+  ospRelease(stfn);
   // Create isosurface geometry //
 
   // create iso geometry object and add it to the world
@@ -102,19 +143,9 @@ int main(int argc, const char **argv)
   // assign material to the geometry
   ospSetObject(isoModel, "material", material);
 
-  // Create slices geometry //
-
-  OSPGeometry sliceGeometry = ospNewGeometry("slices");
-  ospSetObject(sliceGeometry, "volume", volumeModel);
-  float sliceValue = 0.f;
-  setSlice(sliceGeometry, sliceValue);
-  OSPGeometricModel sliceModel = ospNewGeometricModel(sliceGeometry);
-
   // apply changes made
   ospCommit(isoGeometry);
   ospCommit(isoModel);
-  ospCommit(sliceGeometry);
-  ospCommit(sliceModel);
 
   OSPInstance instance = ospNewInstance(group);
   ospCommit(instance);
@@ -134,17 +165,28 @@ int main(int argc, const char **argv)
 
   ospCommit(renderer);
 
+  // create slice geometry
+  float sliceExtent = 1.f;
+  float slicePosition = 0.f;
+  OSPGeometry sliceGeometry = ospNewGeometry("quads");
+  setSliceGeometry(sliceGeometry, sliceExtent, slicePosition);
+  ospCommit(sliceGeometry);
+  OSPGeometricModel sliceModel = ospNewGeometricModel(sliceGeometry);
+
   // Enable show/hide of various objects //
 
   bool showVolume     = true;
   bool showIsoSurface = false;
   bool showSlice      = false;
+  bool changeSliceTransferFunction   = false;
+
 
   auto updateScene = [&]() {
     geometryModelHandles.clear();
 
     ospRemoveParam(group, "geometry");
     ospRemoveParam(group, "volume");
+
 
     if (showVolume) {
       OSPData volumes = ospNewData(1, OSP_OBJECT, &volumeModel);
@@ -156,8 +198,12 @@ int main(int argc, const char **argv)
       if (showIsoSurface)
         geometryModelHandles.push_back(isoModel);
 
-      if (showSlice)
-        geometryModelHandles.push_back(sliceModel);
+      if (showSlice) {
+        if (changeSliceTransferFunction)
+          geometryModelHandles.push_back(setSliceVolumeModel(sliceVolumeModel, sliceModel));
+          else
+            geometryModelHandles.push_back(setSliceVolumeModel(volumeModel, sliceModel));
+      }
 
       OSPData geoms = ospNewData(
           geometryModelHandles.size(), OSP_OBJECT, geometryModelHandles.data());
@@ -202,6 +248,11 @@ int main(int argc, const char **argv)
     if (ImGui::Checkbox("slice", &showSlice))
       updateWorld = true;
 
+    ImGui::NewLine();
+    if (ImGui::Checkbox("change slice TFN", &changeSliceTransferFunction)){
+      updateWorld = true;
+    }
+
     commitWorld = updateWorld;
 
     ImGui::NewLine();
@@ -229,11 +280,10 @@ int main(int argc, const char **argv)
       glfwOSPRayWindow->addObjectToCommit(material);
     }
 
-    if (ImGui::SliderFloat("slice position", &sliceValue, -1.f, 1.f)) {
+    if (ImGui::SliderFloat("slice position", &slicePosition, -1.f, 1.0f)) {
       commitWorld = true;
-      setSlice(sliceGeometry, sliceValue);
+      setSliceGeometry(sliceGeometry, sliceExtent, slicePosition);
       glfwOSPRayWindow->addObjectToCommit(sliceGeometry);
-      glfwOSPRayWindow->addObjectToCommit(sliceModel);
     }
 
     if (updateWorld)
@@ -253,8 +303,9 @@ int main(int argc, const char **argv)
   ospRelease(volumeModel);
   ospRelease(isoModel);
   ospRelease(isoGeometry);
-  ospRelease(sliceGeometry);
   ospRelease(sliceModel);
+  ospRelease(sliceGeometry);
+  ospRelease(sliceVolumeModel);
   ospRelease(material);
   ospRelease(group);
   ospRelease(instance);